bugfix: found undefined params input while implements deletePhoto

+++ src/utils/deletePhoto.ts
@@ -1,4 +1,10 @@
 import fs from "fs";
 export default function deletePhoto(path: string) {
-  fs.unlinkSync(path);
+  if (path !== undefined) {
+    console.log(`-- PATH DETECTED: ${path} --`);
+
+    fs.unlinkSync(path);
+  } else {
+    console.log("-- PATH NOT DETECTED --");
+  }
 }
